+++ src/Framework/Bootstrap/SetupCombinator.php
@@ -28,7 +28,7 @@ final class SetupCombinator
         $this->combineEventDispatcher($other);
         $this->combineServicesToExtend($other);
         $this->combinePlugins($other);
-        $this->combineExtendConfig($other);
+        $this->combineGacelaConfigsToExtend($other);
 
         return $this->original;
     }
@@ -109,10 +109,10 @@ final class SetupCombinator
         }
     }
 
-    private function combineExtendConfig(SetupGacela $other): void
+    private function combineGacelaConfigsToExtend(SetupGacela $other): void
     {
-        if ($other->isPropertyChanged(SetupGacela::extendConfig)) {
-            $this->original->combineExtendConfig($other->getExtendConfig());
+        if ($other->isPropertyChanged(SetupGacela::gacelaConfigsToExtend)) {
+            $this->original->combineGacelaConfigsToExtend($other->getGacelaConfigsToExtend());
         }
     }
 }